saturday night fever

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: samartin <samartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/13 13:28:48 by samartin          #+#    #+#             */
-/*   Updated: 2022/09/15 13:43:00 by samartin         ###   ########.fr       */
+/*   Updated: 2022/09/16 19:45:35 by samartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t n)
 {
-	size_t	i;
+	int	i;
 
 	if (dst > src && dst < (src + n))
 	{
-		i = n - 1;
+		i = (int)n - 1;
 		while (i >= 0)
 		{
 			((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
@@ -28,7 +28,7 @@ void	*ft_memmove(void *dst, const void *src, size_t n)
 	else
 	{
 		i = 0;
-		while (i < n)
+		while (i < (int)n)
 		{
 			((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
 			i++;
